Move pretty-printer FixupContext to a module
Required for being able to make the fields private and force the use of
accessor methods, which will be added in the next commit.
